rocks/ein-agent-worker/pyproject.toml
[project]
name = "ein-agent-worker"
version = "0.1.0"
description = "Temporal worker for Ein Agent AI-powered workflow execution"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dspy>=3.1.2",
    "litellm>=1.80.0",
    "openai-agents>=0.5.1",
    "rich>=14.3.1",
    "temporalio>=1.19.0",
    "typer>=0.21.1",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["ein_agent_worker"]

[project.scripts]
ein-agent-worker = "ein_agent_worker.worker:main_entry"
ein-agent-dspy = "ein_agent_worker.dspy_cli:app"

[dependency-groups]
dev = [
    "temporal-lib-py>=1.8.4",
]
